refactor(repo): remove legacy compatibility layer
