--- apps/web/src/features/admin/lib/options.ts
@@ -1,3 +1,0 @@
-export { adminQueries } from "../api/admin.query";
-export type { AdminUsersQueryInput } from "../api/admin.query";
-export { adminMutations } from "../api/admin.mutation";
