--- apps/web/src/features/anime/lib/options.ts
@@ -1,3 +1,0 @@
-export { animeQueries } from "../api/anime.query";
-export type { AnimeArchiveSearchInput, AnimeSearchInput, HeroCuration } from "../api/anime.query";
-export { animeMutations } from "../api/anime.mutation";
