--- apps/web/src/features/anime/lib/prefetch.server.ts
@@ -1,1 +1,0 @@
-export { prefetchAnimeHome } from "../server/prefetch";
