--- apps/web/src/features/library/lib/options.ts
@@ -1,14 +1,0 @@
-export { LIBRARY_STATUSES, libraryQueries } from "../api/library.query";
-export type { LibraryEntryWithAnime, LibraryStatus } from "../api/library.query";
-export { libraryMutations } from "../api/library.mutation";
-export type {
-  LogAnimeData,
-  LogAnimeInput,
-  RemoveFromLibraryInput,
-  UpdateLibraryProgressData,
-  UpdateLibraryProgressInput,
-  UpdateLibraryRatingData,
-  UpdateLibraryRatingInput,
-  UpdateLibraryStatusData,
-  UpdateLibraryStatusInput,
-} from "../api/library.mutation";
