--- apps/web/src/features/users/lib/options.ts
@@ -1,3 +1,0 @@
-export { userQueries } from "../api/user.query";
-export type { PublicUserLibrary, UpdateProfileData, UserWithProfile } from "../api/user.query";
-export { userMutations } from "../api/user.mutation";
