--- apps/web/src/features/users/lib/prefetch.server.ts
@@ -1,1 +1,0 @@
-export { prefetchProfileLibrary, prefetchUserByUsername } from "../server/prefetch";
