--- packages/api/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@anilog/api",
-  "version": "0.1.0",
-  "type": "module",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "dev": "bun --watch src/index.ts",
-    "build": "bun build src/index.ts",
-    "check-types": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@anilog/config": "workspace:*",
-    "@anilog/contracts": "workspace:*",
-    "@anilog/db": "workspace:*",
-    "@anilog/domain": "workspace:*",
-    "drizzle-orm": "^0.45.1",
-    "elysia": "^1.4.26"
-  },
-  "devDependencies": {
-    "@types/bun": "latest",
-    "typescript": "^5.6.3"
-  }
-}
