--- packages/api/src/errors/api-error.ts
@@ -1,14 +1,0 @@
-export {
-  AppError,
-  ApiError,
-  conflictError,
-  externalServiceError,
-  forbiddenError,
-  internalError,
-  isApiError,
-  notFoundError,
-  toErrorResponse,
-  unauthorizedError,
-  validationError,
-} from "@anilog/domain";
-export type { AppErrorCode } from "@anilog/domain";
