--- packages/api/src/index.ts
@@ -1,6 +1,0 @@
-export * from "./services/anime.service";
-export * from "./services/admin.service";
-export * from "./services/library.service";
-export * from "./services/user.service";
-export * from "./errors/api-error";
-export * from "./schemas";
