--- packages/api/src/schemas/admin.schema.ts
@@ -1,9 +1,0 @@
-export {
-  AdminStatusDto,
-  AdminStatsDto,
-  AdminUsersDto,
-  AdminUsersQuery,
-  SetUserAdminStatusBody,
-  SetUserAdminStatusDto,
-  UpdateHeroCurationBody,
-} from "@anilog/contracts";
