--- packages/api/src/schemas/anime.schema.ts
@@ -1,11 +1,0 @@
-export {
-  AnimeDto,
-  AnimeSearchParams,
-  AnimeSummaryDto,
-  ArchiveSearchDto,
-  ArchiveSearchQuery,
-  HeroCurationDto,
-  HeroCurationParams,
-  UpsertAnimeBody,
-  UpsertAnimeDto,
-} from "@anilog/contracts";
