--- packages/api/src/schemas/common.schema.ts
@@ -1,1 +1,0 @@
-export { SuccessCountDto, UserParams, UsernameParams } from "@anilog/contracts";
