--- packages/api/src/schemas/error.schema.ts
@@ -1,1 +1,0 @@
-export { ErrorCode, ErrorDto, ErrorResponse } from "@anilog/contracts";
