--- packages/api/src/schemas/index.ts
@@ -1,6 +1,0 @@
-export * from "./common.schema";
-export * from "./error.schema";
-export * from "./admin.schema";
-export * from "./anime.schema";
-export * from "./library.schema";
-export * from "./user.schema";
