--- packages/api/src/schemas/library.schema.ts
@@ -1,10 +1,0 @@
-export {
-  LibraryAnimeParams,
-  LibraryEntryDto,
-  LibraryStatusSchema,
-  LogAnimeBody,
-  PublicLibraryEntryDto,
-  UpdateLibraryProgressBody,
-  UpdateLibraryRatingBody,
-  UpdateLibraryStatusBody,
-} from "@anilog/contracts";
