--- packages/api/src/schemas/user.schema.ts
@@ -1,8 +1,0 @@
-export {
-  FollowActionDto,
-  IsFollowingDto,
-  UpdateUserProfileBody,
-  UserProfileDto,
-  UserSearchQuery,
-  UserWithProfileDto,
-} from "@anilog/contracts";
