--- packages/api/src/services/admin.service.ts
@@ -1,1 +1,0 @@
-export { AdminService } from "@anilog/domain";
