--- packages/api/src/services/anime.service.ts
@@ -1,1 +1,0 @@
-export { AnimeService } from "@anilog/domain";
