--- packages/api/src/services/library.service.ts
@@ -1,1 +1,0 @@
-export { LibraryService } from "@anilog/domain";
