--- packages/api/src/services/user.service.ts
@@ -1,1 +1,0 @@
-export { UserService } from "@anilog/domain";
