--- packages/api/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@anilog/config/tsconfig.base.json",
-  "compilerOptions": {
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "outDir": "dist",
-    "composite": true
-  }
-}
